350:IntersectionofTwoArraysII
Using HashMap<K, V> to count and save longer array's element in Map.

Hashmap.getOrDefault<k,v>, Returns the value to which the specified key is mapped, or defaultValue if this map contains no mapping for the key.
HashMap.

HashMap.put<k,v>, Associates the specified value with the specified key in this map. If the map previously contained a mapping for the key, the old value is replaced.

HashMap.remove<k>